checkout prt 2

+++ database/migrations/2020_08_08_001202_create_payments_table.php
@@ -16,7 +16,7 @@ class CreatePaymentsTable extends Migration
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('admin_id')->nullable();
-            $table->decimal('price');
+            $table->double('price', 10,2);
             $table->string('method');
             $table->string('reference')->unique();
             $table->string('image')->nullable();
